# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Documentación Colegio Catolico Api",
    "version": "1.0.0",
    "description": "Documentación de la API de gestor de pagos de Colegio Catolico"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor local"
    }
  ],
  "paths": {
...
    "/padres-familia/crearAcudiente": {
      "post": {
        "summary": "Crea un nuevo acudiente.",
        "tags": [
          "PadreFamilia"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "parentesco": {
                    "type": "string",
                    "description": "Parentesco del acudiente con el estudiante."
                  },
                  "estado": {
                    "type": "string",
                    "description": "Estado del acudiente."
                  },
                  "tipoDocumento": {
                    "type": "string",
                    "description": "Tipo de documento del acudiente."
                  },
                  "identificacion": {
                    "type": "string",
                    "description": "Identificación del acudiente."
                  },
                  "nombres": {
                    "type": "string",
                    "description": "Nombres del acudiente."
                  },
                  "apellidos": {
                    "type": "string",
                    "description": "Apellidos del acudiente."
                  },
                  "profesion": {
                    "type": "string",
                    "description": "Profesión del acudiente."
                  },
                  "dondeTrabaja": {
                    "type": "string",
                    "description": "Lugar de trabajo del acudiente."
                  },
                  "cargo": {
                    "type": "string",
                    "description": "Cargo del acudiente en su trabajo."
                  },
                  "ingresoMensual": {
                    "type": "number",
                    "description": "Ingreso mensual del acudiente."
                  },
                  "correoElectronico": {
                    "type": "string",
                    "description": "Correo electrónico del acudiente."
                  },
                  "direccion": {
                    "type": "string",
                    "description": "Dirección del acudiente."
                  },
                  "telefono": {
                    "type": "string",
                    "description": "Teléfono del acudiente."
                  },
                  "celular": {
                    "type": "string",
                    "description": "Celular del acudiente."
                  },
                  "idEstudiante": {
                    "type": "string",
                    "description": "ID del estudiante asociado."
                  }
                },
                "required": [
                  "parentesco",
                  "estado",
                  "tipoDocumento",
                  "identificacion",
                  "nombres",
                  "apellidos",
                  "profesion",
                  "dondeTrabaja",
                  "cargo",
                  "ingresoMensual",
                  "correoElectronico",
                  "direccion",
                  "telefono",
                  "celular",
                  "idEstudiante"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Acudiente creado exitosamente."
          },
          "400": {
            "description": "Error en los datos enviados."
          }
        }
      }
    },
    "/padres-familia/crearResponsable": {
      "post": {
        "summary": "Crea un nuevo responsable para un estudiante.",
        "tags": [
          "PadreFamilia"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "tipoPersona": {
                    "type": "string",
                    "description": "Tipo de persona (natural o jurídica)."
                  },
                  "razonSocial": {
                    "type": "string",
                    "description": "Razón social del responsable."
                  },
                  "tipoDocumento": {
                    "type": "string",
                    "description": "Tipo de documento del responsable."
                  },
                  "identificacion": {
                    "type": "string",
                    "description": "Número de identificación del responsable."
                  },
                  "direccion": {
                    "type": "string",
                    "description": "Dirección del responsable."
                  },
                  "pais": {
                    "type": "string",
                    "description": "País de residencia del responsable."
                  },
                  "departamento": {
                    "type": "string",
                    "description": "Departamento o estado de residencia del responsable."
                  },
                  "ciudad": {
                    "type": "string",
                    "description": "Ciudad de residencia del responsable."
                  },
                  "correoElectronico": {
                    "type": "string",
                    "description": "Correo electrónico del responsable."
                  },
                  "celular": {
                    "type": "string",
                    "description": "Número de celular del responsable."
                  },
                  "idEstudiante": {
                    "type": "string",
                    "description": "ID del estudiante asociado al responsable."
                  },
                  "responsable": {
                    "type": "string",
                    "description": "Rol del responsable."
                  }
                },
                "required": [
                  "tipoPersona",
                  "razonSocial",
                  "tipoDocumento",
                  "identificacion",
                  "direccion",
                  "pais",
                  "departamento",
                  "ciudad",
                  "correoElectronico",
                  "celular",
                  "idEstudiante",
                  "responsable"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Responsable creado exitosamente."
          },
          "400": {
            "description": "Error en los datos enviados."
          }
        }
      }
    },
    "/padres-familia/listarPadreFamiliaId/{id}": {
      "get": {
        "summary": "Obtiene información de un padre de familia por ID.",
        "tags": [
          "PadreFamilia"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID del padre de familia."
          }
        ],
        "responses": {
          "200": {
            "description": "Información del padre de familia obtenida exitosamente."
          },
          "404": {
            "description": "Padre de familia no encontrado."
          }
        }
      }
    },
    "/padres-familia/getAllAcudiente/{id}": {
      "get": {
        "summary": "Obtiene todos los acudientes asociados a un estudiante por ID.",
        "tags": [
          "PadreFamilia"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID del estudiante."
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de acudientes obtenida exitosamente."
          },
          "404": {
            "description": "Estudiante no encontrado o no tiene acudientes."
          }
        }
      }
    },
...
    "/padres-familia/actualizarAcudiente/{id}": {
      "put": {
        "summary": "Actualiza la información de un acudiente por ID.",
        "tags": [
          "PadreFamilia"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID del acudiente."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "acudiente": {
                    "type": "string",
                    "description": "Nombre del acudiente."
                  },
                  "estudiante": {
                    "type": "string",
                    "description": "ID del estudiante asociado."
                  },
                  "madre": {
                    "type": "string",
                    "description": "Nombre de la madre."
                  },
                  "padre": {
                    "type": "string",
                    "description": "Nombre del padre."
                  },
                  "responsable": {
                    "type": "string",
                    "description": "Nombre del responsable."
                  }
                },
                "required": [
                  "acudiente",
                  "estudiante",
                  "madre",
                  "padre",
                  "responsable"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Información del acudiente actualizada exitosamente."
          },
          "400": {
            "description": "Error en los datos enviados."
          },
          "404": {
            "description": "Acudiente no encontrado."
          }
        }
      }
    },
    "/padres-familia/deshabilitar/{id}": {
      "put": {
        "summary": "Deshabilita un registro por ID.",
        "tags": [
          "PadreFamilia"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID del registro a deshabilitar."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "isActive": {
                    "type": "boolean",
                    "description": "Estado de activación del registro."
                  }
                },
                "required": [
                  "isActive"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Registro deshabilitado exitosamente."
          },
          "400": {
            "description": "Error en los datos enviados."
          },
          "404": {
            "description": "Registro no encontrado."
          }
        }
      }
    },
...
  },
...
}